requirements.txt
SpeechRecognition
pipwin
pyaudio
pyttsx3
newsapi-python